refactor: move predefined nodes out of dns discovery
