+++ src/lib/predefined_bootstrap_nodes.ts
@@ -1,6 +1,6 @@
 import { Multiaddr } from "@multiformats/multiaddr";
 
-import { getPseudoRandomSubset } from "../random_subset";
+import { getPseudoRandomSubset } from "./random_subset";
 
 export const DefaultWantedNumber = 1;
